LayoutTests/fast/events/wheel/first-wheel-event-cancelable-expected.txt

Tests preventDefault on the first event results in the rest of the events being cancelable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS windowScrollEventCount is 0
PASS firstWasCancelable is true
PASS firstWasCanceled is true
PASS secondWasCancelable is true
PASS secondWasCanceled is true
PASS successfullyParsed is true

TEST COMPLETE
